optimization: monitor drawRoundRect cached
