+++ internal/gdi/gdi_canvas.go
@@ -4,15 +4,12 @@
 package gdi
 
 import (
-	"bytes"
 	"image/color"
 	"strings"
 	"unicode"
 	"unicode/utf8"
 
-	svgo "github.com/ajstarks/svgo"
 	"github.com/goplus/canvas"
-	svg "github.com/goplus/spx/internal/svgr"
 	"github.com/hajimehoshi/ebiten/v2"
 )
 
@@ -105,43 +102,6 @@ func DrawLines(target *ebiten.Image, f Font, x, y int, width int, text string, c
 }
 
 // -------------------------------------------------------------------------------------
-
-// Canvas represents a gdi object.
-//
-type Canvas struct {
-	*svgo.SVG
-	Target *ebiten.Image
-}
-
-// Start creates a canvas object.
-//
-func Start(target *ebiten.Image) Canvas {
-	w := new(bytes.Buffer)
-	s := svgo.New(w)
-	cx, cy := target.Size()
-	s.Start(cx, cy)
-	return Canvas{s, target}
-}
-
-// End draws canvas data onto the target image.
-//
-func (p Canvas) End() {
-	p.SVG.End()
-	img, err := svg.Decode(p.SVG.Writer.(*bytes.Buffer))
-	if err != nil {
-		panic(err)
-	}
-	img2 := ebiten.NewImageFromImage(img)
-	defer img2.Dispose()
-	p.Target.DrawImage(img2, nil)
-}
-
-// -------------------------------------------------------------------------------------
-
-type Line struct {
-	text  string
-	width int
-}
 
 type TextMetrics struct {
 	ctx   canvas.Context2D
